Split applications to img and vid.
