+++ README.md
@@ -46,9 +46,10 @@ To use the script, execute the below.
 1. git clone the repository.
 2. Copy `./app/config.sample.yml` to `./app/config.yml`
 3. Open the Makefile and set prompts.
-4. Execute `make app` command. An application will be deployed to Modal.
-5. Execute `make img_by_sdxl_txt2img` command.
-6. Execute `make vid_by_wan_ti2v` command (TI2V video generation).
+4. Execute `make app_img` command. An application for SDXL will be deployed to Modal.
+5. Execute `make app_vid` command. An application for Wan I2V will be deployed to Modal.
+6. Execute `make img_by_sdxl_txt2img` command.
+7. Execute `make vid_by_wan_ti2v` command (TI2V video generation).
 
 Images are generated and output to the `outputs/` directory.
 
@@ -65,7 +66,8 @@ Images are generated and output to the `outputs/` directory.
 │   └── txt2img_handler.py         # A script to run txt2img inference.
 │   └── ti2v_handler.py            # A script to run TI2V inference.
 └── app/                # A directory with config and Modal app.
-    ├── app.py                  # Modal app and inference implementation (SDXL)
+    ├── app_img.py              # Modal app and inference implementation (SDXL)
+    ├── app_vid.py              # Modal app and inference implementation (Wan I2V)
     ├── Dockerfile              # To build a base image.
     ├── config.yml              # To set a model, VAE and optional tools.
     └── requirements.txt
