+++ docs/2026-01-28-ti2v-prd.md
@@ -1,0 +1,97 @@
+# TI2V (Static Image + Text to Video) PRD
+
+## 概要
+
+このPRDは、静止画像とテキストから動画を生成する TI2V 機能を、既存の Modal CLI ベース構成に統合するための要件と設計方針をまとめたもの。
+
+## 目的
+
+- 既存の txt2img CLI と同じ運用体験で TI2V を実行できるようにする。
+- Modal を常時 LISTEN させず、必要時に `modal run` で起動する。
+- 生成動画はローカルに保存し、無料運用を優先する。
+
+## 非目的 (Non-goals)
+
+- Web API の公開・提供は行わない。
+- 生成動画のクラウド配信 (S3 等) は行わない。
+- GUI / WebUI は提供しない。
+
+## 前提・制約
+
+- モデルは `Wan-AI/Wan2.2-TI2V-5B-Diffusers` を使用する。
+- diffusers は main ブランチを使用する (pip 安定版では不足)。
+- 既存の Modal デプロイ (`make app`) を前提とする。
+- 生成は GPU インスタンスで行う (Modal 上)。
+
+## ユーザーフロー
+
+1. `make app` で Modal にアプリをデプロイ。
+2. `make vid_by_wan_ti2v` を実行。
+3. CLI がローカル画像とプロンプトを読み込み、Modal の TI2V クラスを呼び出す。
+4. 生成された mp4 をローカルの `cmd/outputs/` に保存。
+
+## コア仕様
+
+### 入力
+
+- `prompt`: 必須 (空不可)
+- `n_prompt`: 任意
+- `image_path`: 必須 (ローカルの静止画)
+- `height`, `width`: デフォルト 704x1280
+- `steps`: デフォルト 50
+- `num_frames`: デフォルト 121
+- `fps`: デフォルト 24
+- `guidance_scale`: デフォルト 5.0
+- `seed`: -1 の場合はランダム
+- `use_image_aspect`: True/False (画像のアスペクトに合わせる)
+
+### 出力
+
+- `cmd/outputs/YYYY-MM-DD/*.mp4`
+- `prompts_*.txt` に入力パラメータを保存
+
+## アーキテクチャ
+
+### 既存構成に合わせた CLI スタイル
+
+- ローカルエントリポイント: `cmd/ti2v_handler.py`
+- Modal 参照: `modal.Cls.from_name("stable-diffusion-cli", "WanTI2V")`
+- 生成結果は bytes で返し、ローカルに保存
+
+### Modal 側
+
+- `app/app.py` に `WanTI2V` クラスを追加
+- `WanPipeline` + `AutoencoderKLWan` を使用
+- mp4 書き出しは `export_to_video`
+
+## 依存ライブラリ
+
+- torch
+- diffusers (git main)
+- transformers
+- accelerate
+- xformers
+- imageio / imageio-ffmpeg
+- pillow
+
+## コマンド
+
+- デプロイ: `make app`
+- TI2V 実行: `make vid_by_wan_ti2v`
+
+## 運用方針
+
+- API 常時起動はしない
+- モデル重みは Hugging Face から取得
+- 生成物はローカル保存を優先
+
+## 将来の拡張案
+
+- 生成結果が大きくなり運用が辛くなったら Modal Volume 保存に切り替える
+- 必要になれば web endpoint 化を検討
+
+## リスク・注意点
+
+- 動画生成は計算コストが高く、GPU 依存度が高い
+- diffusers main ブランチの変更で挙動が変わる可能性がある
+- ローカル保存はサイズが大きくなりやすい
